update build files and adds Object.assign polyfill.
