+++ lib/react-currency-input.cjs.js
@@ -1,6 +1,26 @@
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
+'use strict';
+
+function _interopDefault (ex) { return (ex && (typeof ex === 'object') && 'default' in ex) ? ex['default'] : ex; }
+
+var PropTypes = _interopDefault(require('prop-types'));
+var React = require('react');
+var React__default = _interopDefault(React);
+var ReactDOM = _interopDefault(require('react-dom'));
+
+Object.assign = Object.assign ||
+  function(target) {
+    var arguments$1 = arguments;
+
+    for (var i = 1; i < arguments.length; i++) {
+      var source = arguments$1[i];
+      for (var key in source) {
+        if (Object.prototype.hasOwnProperty.call(source, key)) {
+          target[key] = source[key];
+        }
+      }
+    }
+    return target;
+  };
 
 function mask(value, precision, decimalSeparator, thousandSeparator, allowNegative, prefix, suffix){
     if ( precision === void 0 ) precision = 2;
@@ -339,13 +359,13 @@ var CurrencyInput = (function (Component$$1) {
         var this$1 = this;
 
         return (
-            React.createElement( 'input', Object.assign({},
+            React__default.createElement( 'input', Object.assign({},
                 { ref: function (input) { this$1.theInput = input; }, type: this.props.inputType, value: this.state.maskedValue, onChange: this.handleChange, onFocus: this.handleFocus, onMouseUp: this.handleFocus }, this.state.customProps))
         )
     };
 
     return CurrencyInput;
-}(Component));
+}(React.Component));
 
 
 
@@ -383,5 +403,5 @@ CurrencyInput.defaultProps = {
     selectAllOnFocus: false
 };
 
-export default CurrencyInput;
-//# sourceMappingURL=redux-zero.es.js.map
+module.exports = CurrencyInput;
+//# sourceMappingURL=react-currency-input.cjs.js.map
