+++ lib/react-currency-input.es.js
@@ -1,11 +1,21 @@
-'use strict';
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 
-function _interopDefault (ex) { return (ex && (typeof ex === 'object') && 'default' in ex) ? ex['default'] : ex; }
+Object.assign = Object.assign ||
+  function(target) {
+    var arguments$1 = arguments;
 
-var PropTypes = _interopDefault(require('prop-types'));
-var React = require('react');
-var React__default = _interopDefault(React);
-var ReactDOM = _interopDefault(require('react-dom'));
+    for (var i = 1; i < arguments.length; i++) {
+      var source = arguments$1[i];
+      for (var key in source) {
+        if (Object.prototype.hasOwnProperty.call(source, key)) {
+          target[key] = source[key];
+        }
+      }
+    }
+    return target;
+  };
 
 function mask(value, precision, decimalSeparator, thousandSeparator, allowNegative, prefix, suffix){
     if ( precision === void 0 ) precision = 2;
@@ -344,13 +354,13 @@ var CurrencyInput = (function (Component$$1) {
         var this$1 = this;
 
         return (
-            React__default.createElement( 'input', Object.assign({},
+            React.createElement( 'input', Object.assign({},
                 { ref: function (input) { this$1.theInput = input; }, type: this.props.inputType, value: this.state.maskedValue, onChange: this.handleChange, onFocus: this.handleFocus, onMouseUp: this.handleFocus }, this.state.customProps))
         )
     };
 
     return CurrencyInput;
-}(React.Component));
+}(Component));
 
 
 
@@ -388,5 +398,5 @@ CurrencyInput.defaultProps = {
     selectAllOnFocus: false
 };
 
-module.exports = CurrencyInput;
-//# sourceMappingURL=redux-zero.cjs.js.map
+export default CurrencyInput;
+//# sourceMappingURL=react-currency-input.es.js.map
