+++ src/index.js
@@ -1,3 +1,5 @@
+import './object-assign-polyfill';
+
 import PropTypes from 'prop-types';
 import React, { Component }  from 'react'
 import ReactDOM from 'react-dom'
